refactor(app): move main logic from Application.kt to Lotto.kt
